教科別の投稿ページに新規投稿機能を実装
