+++ app/controllers/application_controller.rb
@@ -2,6 +2,21 @@ class ApplicationController < ActionController::Base
 
   include SessionsHelper
 
+  # ログイン済みのユーザーかどうか確認
+  def logged_in_user
+    unless logged_in?
+      store_location
+      flash[:danger] = "Please log in."
+      redirect_to login_url, status: :see_other
+    end
+  end
+
+  # 正しいユーザーかどうか確認
+  def correct_user
+    @user = User.find(params[:id])
+    redirect_to(root_url, status: :see_other) unless current_user?(@user)
+  end
+
   def hello
     render html: "テスト成功！！"
   end
