+++ app/controllers/users_controller.rb
@@ -64,19 +64,4 @@ class UsersController < ApplicationController
       @user  = User.find(params[:id])
     end
 
-    # ログイン済みのユーザーかどうか確認
-    def logged_in_user
-      unless logged_in?
-        store_location
-        flash[:danger] = "Please log in."
-        redirect_to login_url, status: :see_other
-      end
-    end
-
-    # 正しいユーザーかどうか確認
-    def correct_user
-      @user = User.find(params[:id])
-      redirect_to(root_url, status: :see_other) unless current_user?(@user)
-    end
-
 end
