+++ app/helpers/subject_posts_helper.rb
@@ -1,0 +1,2 @@
+module SubjectPostsHelper
+end
