+++ app/views/layouts/application.html.erb
@@ -18,6 +18,9 @@
       <%= yield %>
     </div>
   </div>
+  <div class="text-center">
+    <%= debug(params) if Rails.env.development? %>
+  </div>
 </body>
 
 </html>
