+++ app/views/shared/_post_form.html.erb
@@ -1,0 +1,8 @@
+【新規投稿を作成】
+<%= form_with(model: @post) do |f| %>
+  <div><%= f.hidden_field yield(:params_id), :value => params[:id] %></div>
+  <div class="field">
+    <%= f.text_area :content, style: "width: 100%;" %>
+  </div>
+  <%= f.submit "投稿", class: "btn btn-primary" %>
+<% end %>
